luminary/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
google-generativeai
python-multipart
Pillow
pydantic
